Separate `serialize` into its own module
